refactor spam processing file
