+++ src/backend/core/mda/inbound_pipeline.py
@@ -23,7 +23,6 @@ pipeline only sees the uniform ``Step → Decision`` interface.
 from __future__ import annotations
 
 import logging
-import re
 from dataclasses import dataclass, field
 from datetime import timedelta
 from enum import IntEnum
@@ -33,15 +32,14 @@ from django.conf import settings
 from django.db.models import Q
 from django.utils import timezone
 
-import requests
-from jmap_email import JmapEmail, has_header
+from jmap_email import JmapEmail
 
 from core import enums, models
+from core.mda import spam
 from core.mda.inbound_auth import (
     check_inbound_authentication,
     get_inbound_auth_mode,
 )
-from core.mda.utils import headers_blocks
 from core.services.thread_events import assign_users
 
 logger = logging.getLogger(__name__)
@@ -161,155 +159,12 @@ Step = Callable[[InboundContext], Decision]
 # A processing step that keeps failing (a blocking webhook or rspamd today,
 # any future RETRY-returning step tomorrow) must not hold a message
 # forever *or* silently lose it. After this window the task stops holding
-# and delivers the message anyway, stamped with ``X-StMsg-Processing-
-# Failed`` so the UI warns the recipient it bypassed a processing step.
+# and delivers the message anyway, recording ``postmark["processing"]`` so
+# the UI warns the recipient it bypassed a processing step.
 # Generic on purpose — see the RETRY branch in
 # ``process_inbound_message_task``. Operator-tunable via
 # ``MESSAGES_INBOUND_DEFERRAL_MAX_AGE`` (seconds; default 48h).
 DEFERRAL_MAX_AGE = timedelta(seconds=settings.MESSAGES_INBOUND_DEFERRAL_MAX_AGE)
-
-
-# ---------------------------------------------------------------------------
-# Spam-check helpers shared by the steps below.
-# ---------------------------------------------------------------------------
-
-
-def _check_hardcoded_rules(
-    parsed_email: JmapEmail, spam_config: Dict[str, Any]
-) -> Optional[bool]:
-    """Apply the per-domain hardcoded ``rules`` list, header-matched
-    only against headers from trusted relay blocks. Returns ``True`` /
-    ``False`` on first matching rule, ``None`` if no rule matched."""
-    rules = spam_config.get("rules", [])
-    for idx, rule in enumerate(rules):
-        header_match = rule.get("header_match") or rule.get("header_match_regex")
-        if not header_match:
-            continue
-        if ":" not in header_match:
-            # Log the rule position, not its raw value: ``spam_config``
-            # also carries spam-service credentials, so we never echo
-            # values read from it into logs.
-            logger.warning(
-                "Invalid header_match format (missing colon) in spam rule #%d", idx
-            )
-            continue
-        key, value = header_match.split(":", 1)
-        key = key.lower().strip()
-        value = value.lower().strip()
-
-        # Existence check first; the trusted value is read from the
-        # Received-bounded blocks below.
-        if not has_header(parsed_email, key):
-            continue
-
-        # Trust window is "block 0 (our MTA's Received + headers above it)
-        # + N upstream relay blocks". Default 0: a sender can prepend
-        # their own Received lines (landing in block 1+), so trusting
-        # those by default would let them forge an allowlist match.
-        # Slicing beyond list length is fine — yields all blocks.
-        trusted_relays = spam_config.get("trusted_relays", 0)
-        blocks_to_check = trusted_relays + 1
-        found_value = None
-        for block in headers_blocks(parsed_email)[:blocks_to_check]:
-            if key in block and block[key]:
-                # Blocks are ordered most-recent → oldest; first match wins.
-                found_value = block[key][0]
-                break
-        if found_value is None:
-            continue
-
-        header_value = (
-            found_value.lower().strip()
-            if isinstance(found_value, str)
-            else str(found_value).lower().strip()
-        )
-        if rule.get("header_match"):
-            is_match = header_value == value
-        else:  # header_match_regex
-            is_match = re.fullmatch(value, header_value) is not None
-        if is_match:
-            action = rule.get("action") or "spam"
-            if action in ("spam", "reject"):
-                return True
-            if action in ("ham", "no action"):
-                return False
-    return None
-
-
-def _call_rspamd(
-    raw_data: bytes,
-    spam_config: Dict[str, Any],
-    envelope: Optional[Dict[str, Any]] = None,
-) -> Tuple[Optional[str], Optional[str], Optional[Dict[str, Any]]]:
-    """POST raw RFC-822 bytes to rspamd's ``/checkv2``.
-
-    The SMTP ``envelope`` (MAIL FROM, RCPT TO, connecting IP / HELO / rDNS)
-    is forwarded via rspamd's documented scan headers
-    (``From``/``Rcpt``/``IP``/``Helo``/``Hostname``) so envelope-based checks
-    — SPF, DNS RBLs, the Return-Path-vs-From mismatch symbols — score against
-    the real peer. Without them rspamd sees an empty envelope and SPF/RBL
-    degrade to no-ops.
-
-    Returns ``(action, error_message, result_dict)``. ``action`` is the raw
-    rspamd action string (e.g. "no action", "add header", "reject", …) — this
-    function does NOT interpret it; the whole action → verdict/decision mapping
-    lives in a single place, ``_make_rspamd_step``. ``action`` is ``None`` when
-    rspamd is not configured or on error (the ``error_message`` channel
-    distinguishes the two); errors are swallowed so a flaky rspamd never blocks
-    delivery.
-    """
-    url = spam_config.get("rspamd_url")
-    if not url:
-        logger.debug("SPAM_CONFIG.rspamd_url not configured, skipping rspamd")
-        return None, None, None
-
-    headers = {"Content-Type": "message/rfc822"}
-    auth = spam_config.get("rspamd_auth")
-    if auth:
-        headers["Authorization"] = auth
-
-    # Forward the SMTP envelope as rspamd scan headers. Only send the fields
-    # we actually have (widget/internal mail carry no HELO/rDNS); an empty
-    # value would read as a genuinely empty envelope and skew scoring. HELO
-    # and hostname are attacker-influenced, so strip CR/LF to prevent header
-    # injection into the rspamd request.
-    for header_name, envelope_key in (
-        ("From", "mail_from"),
-        ("Rcpt", "rcpt_to"),
-        ("IP", "ip"),
-        ("Helo", "helo"),
-        ("Hostname", "hostname"),
-    ):
-        value = (envelope or {}).get(envelope_key)
-        if value:
-            headers[header_name] = str(value).replace("\r", "").replace("\n", "")
-
-    try:
-        response = requests.post(
-            f"{url}/checkv2", data=raw_data, headers=headers, timeout=10
-        )
-        response.raise_for_status()
-        result = response.json()
-    except (requests.exceptions.RequestException, ValueError) as exc:
-        # Network failures, non-2xx (raise_for_status), and a non-JSON
-        # body (ValueError covers JSONDecodeError) all funnel here. We
-        # don't let a flaky rspamd block delivery — fall through with
-        # is_spam=False so the pipeline keeps moving. The
-        # error_message channel lets the caller log loudly.
-        logger.exception("Error calling rspamd: %s", exc)
-        return None, str(exc), None
-    except Exception as exc:
-        logger.exception("Unexpected error calling rspamd: %s", exc)
-        return None, str(exc), None
-
-    if not isinstance(result, dict):
-        logger.warning("rspamd returned non-object body: %r", result)
-        return None, "rspamd returned non-object body", None
-
-    action = result.get("action", "")
-    score = result.get("score", 0.0)
-    logger.info("Rspamd: action=%s score=%.2f", action, score)
-    return action, None, result
 
 
 # ---------------------------------------------------------------------------
@@ -322,7 +177,7 @@ def _make_hardcoded_rules_step(spam_config: Dict[str, Any]) -> Step:
     def hardcoded_rules(ctx: InboundContext) -> Decision:
         if ctx.is_spam is not None:
             return Decision.CONTINUE
-        verdict = _check_hardcoded_rules(ctx.parsed_email, spam_config)
+        verdict = spam.check_hardcoded_rules(ctx.parsed_email, spam_config)
         if verdict is not None:
             ctx.is_spam = verdict
         return Decision.CONTINUE
@@ -354,7 +209,7 @@ def _make_rspamd_step(spam_config: Dict[str, Any]) -> Step:
     spam-checked. The step RETRYs, so the message is held; if the outage lasts
     past ``DEFERRAL_MAX_AGE`` it is force-delivered flagged rather than
     silently unchecked. (rspamd not being configured is not an error —
-    ``_call_rspamd`` returns no opinion and the pipeline moves on.)
+    ``spam.call_rspamd`` returns no opinion and the pipeline moves on.)
     """
 
     def rspamd(ctx: InboundContext) -> Decision:
@@ -363,7 +218,7 @@ def _make_rspamd_step(spam_config: Dict[str, Any]) -> Step:
             # rspamd's symbols for inbound_auth. The auth step has its
             # own fallback so we can cheaply skip rspamd entirely here.
             return Decision.CONTINUE
-        action, err, result = _call_rspamd(
+        action, err, result = spam.call_rspamd(
             ctx.raw_data, spam_config, envelope=ctx.inbound_message.envelope
         )
         if err:
@@ -428,7 +283,7 @@ def _make_inbound_auth_step(spam_config: Dict[str, Any]) -> Step:
 
     def inbound_auth(ctx: InboundContext) -> Decision:
         if ctx.rspamd_result is None and get_inbound_auth_mode(spam_config) == "rspamd":
-            _, _, ctx.rspamd_result = _call_rspamd(
+            _, _, ctx.rspamd_result = spam.call_rspamd(
                 ctx.raw_data, spam_config, envelope=ctx.inbound_message.envelope
             )
         verdict = check_inbound_authentication(
